requirements.txt
asgiref==3.4.1
boto3==1.19.7
botocore==1.22.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
cryptography==35.0.0
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.8
django-bootstrap-form==3.4
django-filter==21.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
dynaconf==3.1.7
gunicorn==20.1.0
hvac==0.11.2
idna==3.3
Jinja2==3.0.2
jmespath==0.10.0
loguru==0.5.3
MarkupSafe==2.0.1
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
s3transfer==0.5.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
validate-email==1.3
validators==0.18.2
whitenoise==5.3.0
win32-setctime==1.0.3
